+++ server/src-lib/Hasura/RQL/DDL/Permission/Internal.hs
@@ -21,26 +21,10 @@ import           Language.Haskell.TH.Syntax                 (Lift)
 import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.Backends.Postgres.SQL.Value
 import           Hasura.Backends.Postgres.Translate.BoolExp
-import           Hasura.EncJSON
-import           Hasura.Incremental                         (Cacheable)
 import           Hasura.RQL.Types
 import           Hasura.Server.Utils
 import           Hasura.Session
 
-
-data PermColSpec
-  = PCStar
-  | PCCols ![PGCol]
-  deriving (Show, Eq, Lift, Generic)
-instance Cacheable PermColSpec
-
-instance FromJSON PermColSpec where
-  parseJSON (String "*") = return PCStar
-  parseJSON x            = PCCols <$> parseJSON x
-
-instance ToJSON PermColSpec where
-  toJSON (PCCols cols) = toJSON cols
-  toJSON PCStar        = "*"
 
 convColSpec :: FieldInfoMap (FieldInfo 'Postgres) -> PermColSpec -> [PGCol]
 convColSpec _ (PCCols cols) = cols
@@ -134,25 +118,6 @@ dropPermFromCatalog (QualifiedObject sn tn) rn pt =
 
 type CreatePerm a = WithTable (PermDef a)
 
-data PermDef a =
-  PermDef
-  { pdRole       :: !RoleName
-  , pdPermission :: !a
-  , pdComment    :: !(Maybe Text)
-  } deriving (Show, Eq, Lift, Generic)
-instance (Cacheable a) => Cacheable (PermDef a)
-$(deriveFromJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''PermDef)
-
-instance (ToJSON a) => ToJSON (PermDef a) where
-  toJSON = object . toAesonPairs
-
-instance (ToJSON a) => ToAesonPairs (PermDef a) where
- toAesonPairs (PermDef rn perm comment) =
-  [ "role" .= rn
-  , "permission" .= perm
-  , "comment" .= comment
-  ]
-
 data CreatePermP1Res a
   = CreatePermP1Res
   { cprInfo :: !a
@@ -236,61 +201,3 @@ data DropPerm a
 $(deriveJSON (aesonDrop 3 snakeCase){omitNothingFields=True} ''DropPerm)
 
 type family PermInfo a = r | r -> a
-
-class (ToJSON a) => IsPerm a where
-
-  permAccessor
-    :: PermAccessor 'Postgres (PermInfo a)
-
-  buildPermInfo
-    :: (QErrM m, TableCoreInfoRM m)
-    => QualifiedTable
-    -> FieldInfoMap (FieldInfo 'Postgres)
-    -> PermDef a
-    -> m (WithDeps (PermInfo a))
-
-  getPermAcc1
-    :: PermDef a -> PermAccessor 'Postgres (PermInfo a)
-  getPermAcc1 _ = permAccessor
-
-  getPermAcc2
-    :: DropPerm a -> PermAccessor 'Postgres (PermInfo a)
-  getPermAcc2 _ = permAccessor
-
-addPermP2 :: (IsPerm a, MonadTx m, HasSystemDefined m) => QualifiedTable -> PermDef a -> m ()
-addPermP2 tn pd = do
-  let pt = permAccToType $ getPermAcc1 pd
-  systemDefined <- askSystemDefined
-  liftTx $ savePermToCatalog pt tn pd systemDefined
-
-runCreatePerm
-  :: (UserInfoM m, CacheRWM m, IsPerm a, MonadTx m, HasSystemDefined m)
-  => CreatePerm a -> m EncJSON
-runCreatePerm (WithTable tn pd) = do
-  addPermP2 tn pd
-  let pt = permAccToType $ getPermAcc1 pd
-  buildSchemaCacheFor $ MOTableObj tn (MTOPerm (pdRole pd) pt)
-  pure successMsg
-
-dropPermP1
-  :: (QErrM m, CacheRM m, IsPerm a)
-  => DropPerm a -> m (PermInfo a)
-dropPermP1 dp@(DropPerm tn rn) = do
-  tabInfo <- askTabInfo tn
-  askPermInfo tabInfo rn $ getPermAcc2 dp
-
-dropPermP2 :: forall a m. (MonadTx m, IsPerm a) => DropPerm a -> m ()
-dropPermP2 dp@(DropPerm tn rn) =
-  liftTx $ dropPermFromCatalog tn rn pt
-  where
-    pa = getPermAcc2 dp
-    pt = permAccToType pa
-
-runDropPerm
-  :: (IsPerm a, UserInfoM m, CacheRWM m, MonadTx m)
-  => DropPerm a -> m EncJSON
-runDropPerm defn = do
-  dropPermP1 defn
-  dropPermP2 defn
-  withNewInconsistentObjsCheck buildSchemaCache
-  return successMsg
