+++ server/src-lib/Hasura/RQL/DDL/Relationship.hs
@@ -8,7 +8,6 @@ module Hasura.RQL.DDL.Relationship
   , delRelFromCatalog
 
   , runSetRelComment
-  , module Hasura.RQL.DDL.Relationship.Types
   )
 where
 
@@ -26,7 +25,6 @@ import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.EncJSON
 import           Hasura.RQL.DDL.Deps
 import           Hasura.RQL.DDL.Permission          (purgePerm)
-import           Hasura.RQL.DDL.Relationship.Types
 import           Hasura.RQL.Types
 
 runCreateRelationship
@@ -34,7 +32,7 @@ runCreateRelationship
   => RelType -> WithTable (RelDef a) -> m EncJSON
 runCreateRelationship relType (WithTable tableName relDef) = do
   insertRelationshipToCatalog tableName relType relDef
-  buildSchemaCacheFor $ MOTableObj tableName (MTORel (rdName relDef) relType)
+  buildSchemaCacheFor $ MOTableObj tableName (MTORel (_rdName relDef) relType)
   pure successMsg
 
 insertRelationshipToCatalog
@@ -68,7 +66,7 @@ runDropRel (DropRel qt rn cascade) = do
       _       <- askRelType (_tciFieldInfoMap tabInfo) rn ""
       sc      <- askSchemaCache
       let depObjs = getDependentObjs sc (SOTableObj qt $ TORel rn)
-      when (depObjs /= [] && not (or cascade)) $ reportDeps depObjs
+      when (depObjs /= [] && not cascade) $ reportDeps depObjs
       pure depObjs
 
 delRelFromCatalog
