+++ server/src-lib/Hasura/RQL/DDL/Relationship/Rename.hs
@@ -6,7 +6,6 @@ import           Data.Text.Extended
 import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.EncJSON
 import           Hasura.Prelude
-import           Hasura.RQL.DDL.Relationship.Types
 import           Hasura.RQL.DDL.Schema              (renameRelInCatalog)
 import           Hasura.RQL.Types
