+++ server/src-lib/Hasura/RQL/Types/Action.hs
@@ -236,7 +236,7 @@ data ActionPermissionMetadata
 instance NFData ActionPermissionMetadata
 instance Cacheable ActionPermissionMetadata
 
-$(J.deriveFromJSON
+$(J.deriveJSON
   (J.aesonDrop 4 J.snakeCase){J.omitNothingFields=True}
   ''ActionPermissionMetadata)
 
@@ -248,6 +248,7 @@ data ActionMetadata
   , _amDefinition  :: !ActionDefinitionInput
   , _amPermissions :: ![ActionPermissionMetadata]
   } deriving (Show, Eq, Lift, Generic)
+$(J.deriveToJSON (J.aesonDrop 3 J.snakeCase) ''ActionMetadata)
 instance NFData ActionMetadata
 instance Cacheable ActionMetadata
