+++ server/src-lib/Hasura/RQL/Types/Common.hs
@@ -48,6 +48,10 @@ module Hasura.RQL.Types.Common
        , unsafeNonNegativeInt
        , Timeout(..)
        , defaultActionTimeoutSecs
+
+       , UrlConf(..)
+       , resolveUrlConf
+       , getEnv
        ) where
 
 import           Hasura.Prelude
@@ -99,7 +103,7 @@ rootText = mkNonEmptyTextUnsafe "root"
 
 newtype RelName
   = RelName { getRelTxt :: NonEmptyText }
-  deriving (Show, Eq, Hashable, FromJSON, ToJSON, Q.ToPrepArg, Q.FromCol, Lift, Generic, Arbitrary, NFData, Cacheable)
+  deriving (Show, Eq, Hashable, FromJSON, ToJSON, ToJSONKey, Q.ToPrepArg, Q.FromCol, Lift, Generic, Arbitrary, NFData, Cacheable)
 
 instance IsIdentifier RelName where
   toIdentifier rn = Identifier $ relNameToTxt rn
@@ -351,3 +355,35 @@ instance Arbitrary Timeout where
 
 defaultActionTimeoutSecs :: Timeout
 defaultActionTimeoutSecs = Timeout 30
+
+data UrlConf
+  = UrlValue !InputWebhook
+  | UrlFromEnv !T.Text
+  deriving (Show, Eq, Generic, Lift)
+instance NFData UrlConf
+instance Cacheable UrlConf
+
+instance ToJSON UrlConf where
+  toJSON (UrlValue w)      = toJSON w
+  toJSON (UrlFromEnv wEnv) = object ["from_env" .= wEnv ]
+
+instance FromJSON UrlConf where
+  parseJSON (Object o) = UrlFromEnv <$> o .: "from_env"
+  parseJSON t@(String _) =
+    case (fromJSON t) of
+      Error s   -> fail s
+      Success a -> pure $ UrlValue a
+  parseJSON _          = fail "one of string or object must be provided for url/webhook"
+
+resolveUrlConf
+  :: MonadError QErr m => Env.Environment -> UrlConf -> m Text
+resolveUrlConf env = \case
+  UrlValue v        -> unResolvedWebhook <$> resolveWebhook env v
+  UrlFromEnv envVar -> getEnv env envVar
+
+getEnv :: QErrM m => Env.Environment -> T.Text -> m T.Text
+getEnv env k = do
+  let mEnv = Env.lookupEnv env (T.unpack k)
+  case mEnv of
+    Nothing     -> throw400 NotFound $ "environment variable '" <> k <> "' not set"
+    Just envVal -> return (T.pack envVal)
