+++ server/src-lib/Hasura/RQL/Types/Relationship.hs
@@ -1,26 +1,29 @@
-module Hasura.RQL.DDL.Relationship.Types where
+module Hasura.RQL.Types.Relationship where
 
+import           Hasura.Backends.Postgres.SQL.Types
+import           Hasura.Incremental                 (Cacheable)
 import           Hasura.Prelude
 
-import qualified Data.HashMap.Strict                as HM
-
+import           Control.Lens                       (makeLenses)
 import           Data.Aeson.Casing
 import           Data.Aeson.TH
 import           Data.Aeson.Types
+import           Hasura.RQL.Types.Common
 import           Instances.TH.Lift                  ()
 import           Language.Haskell.TH.Syntax         (Lift)
 
-import           Hasura.Backends.Postgres.SQL.Types
-import           Hasura.RQL.Types
+import qualified Data.HashMap.Strict                as HM
+import qualified Data.Text                          as T
 
 data RelDef a
   = RelDef
-  { rdName    :: !RelName
-  , rdUsing   :: !a
-  , rdComment :: !(Maybe Text)
+  { _rdName    :: !RelName
+  , _rdUsing   :: !a
+  , _rdComment :: !(Maybe T.Text)
   } deriving (Show, Eq, Lift, Generic)
-
-$(deriveFromJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''RelDef)
+instance (Cacheable a) => Cacheable (RelDef a)
+$(deriveFromJSON (aesonDrop 3 snakeCase){omitNothingFields=True} ''RelDef)
+$(makeLenses ''RelDef)
 
 instance (ToJSON a) => ToJSON (RelDef a) where
   toJSON = object . toAesonPairs
@@ -37,6 +40,7 @@ data RelManualConfig
   { rmTable   :: !QualifiedTable
   , rmColumns :: !(HashMap PGCol PGCol)
   } deriving (Show, Eq, Lift, Generic)
+instance Cacheable RelManualConfig
 
 instance FromJSON RelManualConfig where
   parseJSON (Object v) =
@@ -57,6 +61,7 @@ data RelUsing a
   = RUFKeyOn !a
   | RUManual !RelManualConfig
   deriving (Show, Eq, Lift, Generic)
+instance (Cacheable a) => Cacheable (RelUsing a)
 
 instance (ToJSON a) => ToJSON (RelUsing a) where
   toJSON (RUFKeyOn fkey) =
@@ -82,6 +87,7 @@ data ArrRelUsingFKeyOn
   { arufTable  :: !QualifiedTable
   , arufColumn :: !PGCol
   } deriving (Show, Eq, Lift, Generic)
+instance Cacheable ArrRelUsingFKeyOn
 
 $(deriveJSON (aesonDrop 4 snakeCase){omitNothingFields=True} ''ArrRelUsingFKeyOn)
 
@@ -97,19 +103,30 @@ data DropRel
   = DropRel
   { drTable        :: !QualifiedTable
   , drRelationship :: !RelName
-  , drCascade      :: !(Maybe Bool)
+  , drCascade      :: !Bool
   } deriving (Show, Eq, Lift)
+$(deriveToJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''DropRel)
 
-$(deriveJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''DropRel)
+instance FromJSON DropRel where
+  parseJSON = withObject "Object" $ \o ->
+    DropRel
+      <$> o .: "table"
+      <*> o .: "relationship"
+      <*> o .:? "cascade" .!= False
 
 data SetRelComment
   = SetRelComment
   { arTable        :: !QualifiedTable
   , arRelationship :: !RelName
-  , arComment      :: !(Maybe Text)
+  , arComment      :: !(Maybe T.Text)
   } deriving (Show, Eq, Lift)
-
-$(deriveJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''SetRelComment)
+$(deriveToJSON (aesonDrop 2 snakeCase){omitNothingFields=True} ''SetRelComment)
+instance FromJSON SetRelComment where
+  parseJSON = withObject "Object" $ \o ->
+    SetRelComment
+      <$> o .: "table"
+      <*> o .: "relationship"
+      <*> o .:? "comment"
 
 data RenameRel
   = RenameRel
@@ -117,5 +134,11 @@ data RenameRel
   , rrName    :: !RelName
   , rrNewName :: !RelName
   } deriving (Show, Eq, Lift)
+$(deriveToJSON (aesonDrop 2 snakeCase) ''RenameRel)
 
-$(deriveJSON (aesonDrop 2 snakeCase) ''RenameRel)
+instance FromJSON RenameRel where
+  parseJSON = withObject "Object" $ \o ->
+    RenameRel
+      <$> o .: "table"
+      <*> o .: "name"
+      <*> o .: "new_name"
