+++ server/src-lib/Hasura/Server/API/Query.hs
@@ -115,7 +115,7 @@ data RQLQueryV1
 
   | RQRunSql !RunSQL
 
-  | RQReplaceMetadata !ReplaceMetadata
+  | RQReplaceMetadata !Metadata
   | RQExportMetadata !ExportMetadata
   | RQClearMetadata !ClearMetadata
   | RQReloadMetadata !ReloadMetadata
